requirements.txt
connectrum==0.8.1
python-bitcoinrpc==1.0
